r2 added view engine and simply display the products data
the html files are still there operating separately to emulate remote request. Whereas, the pug views are to display raw data for the purpose of rapid testing and development. With time, an additional layer of testing written in Mocha, Chai, Protractor,... can be provided to perform better assert() and describe(), incorporating testing automation and affirming the build result